[#93] move: enum 폴더 옮기기
