+++ src/main/java/com/beour/space/domain/enums/SpaceCategory.java
@@ -1,4 +1,4 @@
-package com.beour.space.host.enums;
+package com.beour.space.domain.enums;
 
 public enum SpaceCategory {
     CAFE, RESTAURANT, COOKING, LEATHER, COSTUME, ART
