+++ src/main/java/com/beour/space/guest/controller/GuestSpaceController.java
@@ -8,8 +8,8 @@ import com.beour.space.guest.dto.RecentCreatedSpcaceListResponseDto;
 import com.beour.space.guest.dto.SearchSpaceResponseDto;
 import com.beour.space.guest.service.GuestSpaceSearchService;
 import com.beour.space.guest.service.GuestSpaceService;
-import com.beour.space.host.enums.SpaceCategory;
-import com.beour.space.host.enums.UseCategory;
+import com.beour.space.domain.enums.SpaceCategory;
+import com.beour.space.domain.enums.UseCategory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
