+++ src/main/java/com/beour/space/guest/service/GuestSpaceSearchService.java
@@ -6,8 +6,8 @@ import com.beour.space.domain.entity.Space;
 import com.beour.space.domain.repository.SpaceRepository;
 import com.beour.space.guest.dto.FilteringSearchRequestDto;
 import com.beour.space.guest.dto.SearchSpaceResponseDto;
-import com.beour.space.host.enums.SpaceCategory;
-import com.beour.space.host.enums.UseCategory;
+import com.beour.space.domain.enums.SpaceCategory;
+import com.beour.space.domain.enums.UseCategory;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
