+++ src/test/java/com/beour/reservation/guest/service/CheckAvailableTimeServiceTest.java
@@ -14,8 +14,8 @@ import com.beour.space.domain.entity.Space;
 import com.beour.space.domain.repository.AvailableTimeRepository;
 import com.beour.space.domain.repository.SpaceRepository;
 import com.beour.space.domain.repository.TagRepository;
-import com.beour.space.host.enums.SpaceCategory;
-import com.beour.space.host.enums.UseCategory;
+import com.beour.space.domain.enums.SpaceCategory;
+import com.beour.space.domain.enums.UseCategory;
 import com.beour.user.entity.User;
 import com.beour.user.repository.UserRepository;
 import java.time.LocalDate;
