+++ src/test/java/com/beour/wishlist/service/WishlistServiceTest.java
@@ -10,8 +10,8 @@ import com.beour.space.domain.entity.Tag;
 import com.beour.space.domain.repository.SpaceRepository;
 import com.beour.space.domain.repository.TagRepository;
 import com.beour.space.guest.dto.SpaceListSpaceResponseDto;
-import com.beour.space.host.enums.SpaceCategory;
-import com.beour.space.host.enums.UseCategory;
+import com.beour.space.domain.enums.SpaceCategory;
+import com.beour.space.domain.enums.UseCategory;
 import com.beour.user.entity.User;
 import com.beour.user.repository.UserRepository;
 import com.beour.wishlist.entity.Like;
